Render alerts on action success

+++ client/components/adminComponents/role-card.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Card from '@material-ui/core/Card';
 import { makeStyles } from '@material-ui/core/styles';
+import Swal from 'sweetalert2';
 import CardContent from '@material-ui/core/CardContent';
 import DeleteIcon from '@material-ui/icons/Delete';
 import Tooltip from '@material-ui/core/Tooltip';
@@ -11,6 +12,14 @@ import { formatDate } from '../../utils';
 import { deleteRole } from '../../actions';
 
 const useStyles = makeStyles({
+    root: {
+        backgroundColor: '#CDCDCD'
+    },
+    header: {
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center'
+    },
     delete: {
         height: '1.5rem',
         width: '2rem',
@@ -27,12 +36,30 @@ const RoleCard = ({ role, deleteRole }) => {
     const classes = useStyles();
 
     const destroyRole = roleId => {
-        deleteRole(roleId);
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: 'rgb(80, 143, 202)',
+            cancelButtonColor: '#8B0000',
+            confirmButtonText: 'Yes, delete it!'
+        }).then(async result => {
+            if (result.value) {
+                await deleteRole(roleId);
+                Swal.fire({
+                    title: 'Role deleted',
+                    icon: 'success',
+                    showConfirmButton: false,
+                    timer: 1500
+                });
+            }
+        });
     };
 
     return (
-        <Card className="role-card">
-            <CardContent className="role-card__header">
+        <Card className={classes.root}>
+            <CardContent className={classes.header}>
                 <Typography className="role-title__wrapper">
                     title: <span className="role-title">{role.title}</span>
                 </Typography>
